added update telegram

+++ src/const/CardPageData.js
@@ -1,7 +1,9 @@
 const en = {
     name: 'Name',
+    condition: 'Condition',
     price: 'Price',
     city: 'City',
+    addDescription: 'Add description',
     seeAllUserAds: 'See all user ads',
     deleteBtn: 'Delete ad',
     deleteFromFavBtn: 'Delete from favorites',
@@ -13,8 +15,10 @@ const en = {
 
 const rus = {
     name: 'Название',
+    condition: 'Состояние',
     price: 'Цена',
     city: 'Город',
+    addDescription: 'Добавить описание',
     seeAllUserAds: 'Посмотреть все объявления',
     deleteBtn: 'Удалить',
     deleteFromFavBtn: 'Удалить из избранного',
@@ -26,8 +30,10 @@ const rus = {
 
 const hebrew = {
     name: 'Name',
+    condition: 'Condition',
     price: 'Price',
     city: 'City',
+    addDescription: 'Add description',
     seeAllUserAds: 'See all user ads',
     deleteBtn: 'Delete ad',
     deleteFromFavBtn: 'Delete from favorites',
